+++ crates/holochain/CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+- Fix an issue which caused a `WebsocketClosed` error to be logged when signals are sent after a client has closed their
+  connection.
+
 ## 0.2.8
 
 ## 0.2.8-rc.1